% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{add_n_pixels}
\alias{add_n_pixels}
\title{Add pixel information to extracted data}
\usage{
add_n_pixels(roi_df, bm_r, roi_sf, quiet)
}
\arguments{
\item{roi_df}{Data frame containing extracted data.}

\item{bm_r}{The raster data.}

\item{roi_sf}{The spatial features representing the regions of interest.}

\item{quiet}{Logical indicating whether to show progress messages.}
}
\value{
The input data frame with added columns for pixel information.
}
\description{
This function calculates additional pixel information for the extracted data frames.
}
\examples{
# Adding pixel information to extracted data
add_n_pixels(roi_df, bm_r, roi_sf, quiet = FALSE)

}
