% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{apply_scaling_factor_to_viirs_data}
\alias{apply_scaling_factor_to_viirs_data}
\title{Apply Scaling Factor to VIIRS Data}
\usage{
apply_scaling_factor_to_viirs_data(x, variable, quiet = TRUE)
}
\arguments{
\item{x}{A numeric vector or matrix representing the VIIRS data.}

\item{variable}{A character string specifying the variable name.}

\item{quiet}{Logical, if TRUE, suppresses warnings when the variable is not found in scaling variables.}
}
\value{
A numeric vector or matrix with the scaling factor applied to the specified variables.
}
\description{
Apply scaling factor to variables according to Black Marble user guide.
The scaling factor is 0.1 for specific VIIRS variables.
}
\details{
This function applies a scaling factor of 0.1 to specific VIIRS variables
according to the Black Marble user guide.

The following VIIRS variables are affected:
\itemize{
\item DNB_At_Sensor_Radiance (VNP46A1)
\item DNB_BRDF-Corrected_NTL (VNP46A2)
\item Gap_Filled_DNB_BRDF-Corrected_NTL (VNP46A2)
\item DNB_Lunar_Irradiance (VNP46A2)
\item AllAngle_Composite_Snow_Covered (VNP46A3/4)
\item AllAngle_Composite_Snow_Covered_Std (VNP46A3/4)
\item AllAngle_Composite_Snow_Free (VNP46A3/4)
\item AllAngle_Composite_Snow_Free_Std (VNP46A3/4)
\item NearNadir_Composite_Snow_Covered (VNP46A3/4)
\item NearNadir_Composite_Snow_Covered_Std (VNP46A3/4)
\item NearNadir_Composite_Snow_Free (VNP46A3/4)
\item NearNadir_Composite_Snow_Free_Std (VNP46A3/4)
\item OffNadir_Composite_Snow_Covered (VNP46A3/4)
\item OffNadir_Composite_Snow_Covered_Std (VNP46A3/4)
\item OffNadir_Composite_Snow_Free (VNP46A3/4)
\item OffNadir_Composite_Snow_Free_Std (VNP46A3/4)
}
}
\references{
Black Marble User Guide - https://viirsland.gsfc.nasa.gov/PDF/BlackMarbleUserGuide_v1.2_20220916.pdf
}
