% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{bind_extracted_data}
\alias{bind_extracted_data}
\title{Bind extracted data}
\usage{
bind_extracted_data(extracted_data_list)
}
\arguments{
\item{...}{One or more extracted data frames or a list of extracted data frames.}
}
\value{
A single data frame containing the bound extracted data.
}
\description{
This function binds together the extracted data frames or a list of extracted data frames.
}
\examples{
# Binding extracted data frames
bind_extracted_data(df1, df2)

# Binding a list of extracted data frames
bind_extracted_data(list_of_dfs)
}
