% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarbler.R
\name{bm_extract}
\alias{bm_extract}
\title{Extracts raster data for specified region of interest}
\usage{
bm_extract(
  roi_sf,
  product_id,
  date,
  bearer,
  aggregation_fun = c("mean"),
  add_n_pixels = TRUE,
  variable = NULL,
  quality_flags_to_remove = NULL,
  check_all_tiles_exist = TRUE,
  interpol_na = FALSE,
  output_location_type = "memory",
  file_dir = NULL,
  file_prefix = NULL,
  file_skip_if_exists = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{roi_sf}{Spatial object representing the region of interest.}

\item{product_id}{Identifier for the product.}

\item{date}{Dates for which raster data is to be extracted.}

\item{bearer}{Authentication token for accessing the data.}

\item{aggregation_fun}{Aggregation function to summarize raster values (default is "mean").}

\item{add_n_pixels}{Logical indicating whether to add pixel count information (default is TRUE).}

\item{variable}{Variable to extract from the product (default is NULL).}

\item{quality_flags_to_remove}{Quality flag removal options (default is NULL).}

\item{check_all_tiles_exist}{Logical indicating whether to check if all tiles exist (default is TRUE).}

\item{interpol_na}{Logical indicating whether to interpolate missing values (default is FALSE).}

\item{output_location_type}{Output location type ("memory" or "file", default is "memory").}

\item{file_dir}{Directory for saving output files (used if output_location_type is "file", default is NULL).}

\item{file_prefix}{Prefix for output file names (default is NULL).}

\item{file_skip_if_exists}{Logical indicating whether to skip file creation if output file already exists (default is TRUE).}

\item{quiet}{Logical indicating whether to suppress progress messages (default is FALSE).}
}
\value{
A data frame containing extracted raster data.
}
\description{
Extracts raster data for specified region of interest
}
