% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarbler.R
\name{bm_raster}
\alias{bm_raster}
\title{Make Black Marble Raster}
\usage{
bm_raster(roi_sf, product_id, date, bearer, variable = NULL)
}
\arguments{
\item{roi_sf}{Region of interest; sf polygon. Must be in the \href{https://epsg.io/4326}{WGS 84 (epsg:4326)} coordinate reference system.}

\item{product_id}{Either: \code{VNP46A1}, \code{VNP46A2}, \code{VNP46A3}, or \code{VNP46A4}.}

\item{date}{Date of raster data. For \code{VNP46A1} and \code{VNP46A2} (daily data), a date (eg, \code{"2021-10-03"}). For \code{VNP46A3} (monthly data), a date or year-month (e.g., (a) \code{"2021-10-01"}, where the day will be ignored, or (b) \code{"2021-10"}). For \code{VNP46A4} (annual data), year or date  (e.g., (a) \code{"2021-10-01"}, where the month and day will be ignored, or (b) \code{"2021"}). Entering one date will produce a raster. Entering multiple dates will produce a raster stack.}

\item{bearer}{NASA bearer token.}

\item{variable}{Variable to used to create raster (default: \code{NULL}). If \code{NULL}, uses a a default variable. For \code{VNP46A1}, uses \code{DNB_At_Sensor_Radiance_500m}. For \code{VNP46A2}, uses \code{Gap_Filled_DNB_BRDF-Corrected_NTL}. For \code{VNP46A3} and \code{VNP46A4}, uses \code{NearNadir_Composite_Snow_Free}. For information on other variable choices, see \href{https://ladsweb.modaps.eosdis.nasa.gov/api/v2/content/archives/Document\%20Archive/Science\%20Data\%20Product\%20Documentation/VIIRS_Black_Marble_UG_v1.2_April_2021.pdf}{here}; for \code{VNP46A1}, see Table 3; for \code{VNP46A2} see Table 6; for \code{VNP46A3} and \code{VNP46A4}, see Table 9.}
}
\value{
Raster
}
\description{
Make a raster of nighttime lights from \href{https://blackmarble.gsfc.nasa.gov/}{NASA Black Marble data}
}
\examples{
\dontrun{
# Define bearer token
bearer <- "BEARER-TOKEN-HERE"

# sf polygon of Kenya
roi_sf <- getData('GADM', country='KEN', level=0) \%>\% st_as_sf()

# Daily data: raster for October 3, 2021
ken_20210205_r <- bm_raster(roi_sf = roi_sf,
                            product_id = "VNP46A2",
                            date = "2021-10-03",
                            bearer = bearer)

# Monthly data: raster for March 2021
ken_202103_r <- bm_raster(roi_sf = roi_sf,
                          product_id = "VNP46A3",
                          date = "2021-03-01",
                          bearer = bearer)

# Annual data: raster for 2021
ken_2021_r <- bm_raster(roi_sf = roi_sf,
                        product_id = "VNP46A4",
                        date = 2021,
                        bearer = bearer)
}

}
