% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarbler.R
\docType{data}
\name{bm_tiles_sf}
\alias{bm_tiles_sf}
\title{Black Marble Tile Grid Shapefile
A dataset containing black marble grid tiles.}
\format{
Sf polygon with 648 features
}
\usage{
data(bm_tiles_sf)
}
\description{
Black Marble Tile Grid Shapefile
A dataset containing black marble grid tiles.
}
\references{
\url{https://blackmarble.gsfc.nasa.gov/}
}
\author{
Robert Marty \email{rmarty@worldbank.org}
}
\keyword{datasets}
