% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{clean_raster_data}
\alias{clean_raster_data}
\title{Clean Raster Data}
\usage{
clean_raster_data(raster_obj, variable_name)
}
\arguments{
\item{raster_obj}{The raster object to be cleaned.}

\item{variable_name}{A character string specifying the variable name.}
}
\value{
A cleaned raster object.
}
\description{
This function cleans raster data by removing fill values and applying scaling factors.
}
\details{
This function removes fill values and applies scaling factors to the specified variable in the raster object.
}
\seealso{
\link{remove_fill_value_from_satellite_data}, \link{apply_scaling_factor_to_viirs_data}
}
