% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{convert_h5_to_raster}
\alias{convert_h5_to_raster}
\title{Convert HDF5 File to Raster}
\usage{
convert_h5_to_raster(
  download_path,
  blackmarble_variable,
  quality_flags_to_remove = numeric()
)
}
\arguments{
\item{blackmarble_variable}{A character string specifying the blackmarble_variable name to extract from the HDF5 file.}

\item{quality_flags_to_remove}{A numeric vector containing quality flag values to be removed from the data (optional).}

\item{file_path}{A character string representing the filepath to the HDF5 file.}
}
\value{
A raster object containing the extracted variable data from the HDF5 file.
}
\description{
Converts an HDF5 file to a raster object.
}
\details{
This function converts an HDF5 file to a raster object. It extracts the specified variable
from the HDF5 file and optionally removes specific quality flag values from the data.
}
\references{
Black Marble User Guide - https://viirsland.gsfc.nasa.gov/PDF/BlackMarbleUserGuide_v1.2_20220916.pdf
}
