% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{create_black_marble_dataset_df}
\alias{create_black_marble_dataset_df}
\title{Create Black Marble Dataset DataFrame}
\usage{
create_black_marble_dataset_df(
  product_id,
  all = TRUE,
  years = NULL,
  months = NULL,
  days = NULL
)
}
\arguments{
\item{product_id}{The product ID specifying the type of Black Marble data.}

\item{all}{Logical; indicating whether to create filenames for all available data or only for specific years, months, or days (default: TRUE).}

\item{years}{A numeric vector specifying the years for which to create filenames (optional).}

\item{months}{A numeric vector specifying the months for which to create filenames (optional).}

\item{days}{A numeric vector specifying the days for which to create filenames (optional).}
}
\value{
A data frame containing the filenames of Black Marble datasets based on the specified parameters.
}
\description{
Creates a data frame containing Black Marble dataset filenames based on the specified parameters.
}
\details{
This function generates a data frame with filenames of Black Marble datasets. It allows filtering
by year, month, and day based on the provided parameters. Depending on the \code{product_id}, it generates
filenames for daily, monthly, or yearly data. The generated filenames are based on the specified product ID,
year, and day or month. The resulting data frame includes columns for the year and day or month, depending on the
type of data.
}
\examples{
\dontrun{
# Generate filenames for all available Black Marble data
all_data <- create_black_marble_dataset_df("VNP46A1")

# Generate filenames for Black Marble data for specific years and months
specific_data <- create_black_marble_dataset_df("VNP46A2", years = c(2018, 2019), months = 1:6)
}
}
