% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{create_raster_from_data_metadata}
\alias{create_raster_from_data_metadata}
\title{Create Raster Object from Data and Metadata using terra}
\usage{
create_raster_from_data_metadata(data, metadata)
}
\arguments{
\item{data}{The data matrix for the raster.}

\item{metadata}{A list containing metadata information including nodata value, bounding box coordinates, and coordinate reference system (CRS).}
}
\value{
A raster object created from the input data and metadata.
}
\description{
This function creates a raster object from the provided data and metadata using the terra package.
}
\details{
This function transposes the input data, assigns nodata values to NA, creates an extent class from bounding box coordinates, and finally creates a raster object with the given data, extent, and CRS.
}
