% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{custom_inform}
\alias{custom_inform}
\title{Custom Inform Function}
\usage{
custom_inform(message, quiet = FALSE)
}
\arguments{
\item{message}{The message to be printed.}

\item{quiet}{Logical indicating whether to suppress printing the message. Default is FALSE.}
}
\description{
This function prints an information message if quiet is set to FALSE.
}
\examples{
\dontrun{
custom_inform("This is an information message.", quiet = FALSE)
custom_inform("This message will not be printed.", quiet = TRUE)
}
}
