% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{define_raster_name}
\alias{define_raster_name}
\title{Define Raster Name by Date}
\usage{
define_raster_name(date_string, product_id)
}
\arguments{
\item{date_string}{A character string representing the date in the format "YYYY-MM-DD".}

\item{product_id}{A character string representing the product ID (e.g., "VNP46A1", "VNP46A2").}
}
\value{
A character string representing the generated raster name.
}
\description{
Generates a name for the raster based on the given date and product ID.
}
