% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{download_and_convert_raster}
\alias{download_and_convert_raster}
\title{Download and Convert Raster Data}
\usage{
download_and_convert_raster(
  file_name,
  temp_dir,
  variable,
  bearer,
  quality_flags_to_remove = numeric(),
  quiet = FALSE
)
}
\arguments{
\item{file_name}{A character string representing the name of the file to download.}

\item{temp_dir}{A character string specifying the temporary directory where the file will be saved.}

\item{variable}{A character string specifying the variable to extract from the raster data.}

\item{bearer}{A character string containing the authorization token for accessing NASA's LADSWeb.}

\item{quality_flags_to_remove}{A numeric vector containing quality flag values to be removed from the data (optional).}

\item{quiet}{Logical; indicating whether to suppress progress messages (default: FALSE).}
}
\value{
A raster object containing the downloaded and processed raster data.
}
\description{
Downloads raster data from NASA's LADSWeb and converts it to a raster object.
}
\details{
This function downloads raster data from NASA's LADSWeb based on the provided file name.
It then converts the downloaded data to a raster object, extracting the specified variable and removing
quality flag values if specified. The function also provides an option to suppress progress messages.
}
\references{
Black Marble User Guide - https://viirsland.gsfc.nasa.gov/PDF/BlackMarbleUserGuide_v1.2_20220916.pdf
}
