% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{download_h5_viirs_sat_image}
\alias{download_h5_viirs_sat_image}
\title{Download VIIRS Satellite Image in HDF5 Format}
\usage{
download_h5_viirs_sat_image(
  file_name,
  download_path,
  bearer,
  quality_flags_to_remove = numeric(),
  quiet = FALSE
)
}
\arguments{
\item{file_name}{A character string representing the name of the file to download.}

\item{download_path}{A character string representing the path to save the downloaded file.}

\item{bearer}{A character string containing the authorization token for accessing NASA's LADSWeb.}

\item{quality_flags_to_remove}{A numeric vector containing quality flag values to be removed from the data (optional).}

\item{quiet}{Logical; indicating whether to suppress progress messages (default: FALSE).}
}
\description{
Downloads a VIIRS satellite image in HDF5 format from NASA's LADSWeb and saves it to a temporary directory or persistent location.
}
\details{
This function downloads a VIIRS satellite image in HDF5 format from NASA's LADSWeb
based on the provided file name. It then processes the downloaded data, removing quality flag values
if specified. The function also provides an option to suppress progress messages.
}
\references{
Black Marble User Guide - https://viirsland.gsfc.nasa.gov/PDF/BlackMarbleUserGuide_v1.2_20220916.pdf
}
