% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{extract_and_process}
\alias{extract_and_process}
\title{Extract and process raster data}
\usage{
extract_and_process(raster, roi_sf, fun, add_n_pixels = TRUE, quiet)
}
\description{
Extract and process raster data
}
