% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{extract_and_process_i}
\alias{extract_and_process_i}
\title{Extract and process raster data for individual dates}
\usage{
extract_and_process_i(
  roi_sf,
  product_id,
  date_i,
  bearer,
  variable,
  quality_flags_to_remove,
  check_all_tiles_exist,
  add_n_pixels = TRUE,
  quiet,
  temp_dir
)
}
\description{
Extract and process raster data for individual dates
}
