% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{extract_bounding_box}
\alias{extract_bounding_box}
\title{Extracts bounding box coordinates based on tile ID from the file path}
\usage{
extract_bounding_box(file_path, black_marble_tiles_sf)
}
\arguments{
\item{file_path}{A character string specifying the file path.}

\item{black_marble_tiles_sf}{The spatial dataset containing the grid tiles information.}
}
\value{
A list containing the minimum and maximum longitude and latitude coordinates of the bounding box.
}
\description{
This code snippet extracts the bounding box coordinates (minimum longitude, minimum latitude, maximum longitude, maximum latitude)
based on the tile ID extracted from the file path. It uses the \code{file_path} to extract the tile ID, then retrieves the corresponding
grid from the \code{black_marble_tiles_sf} dataset. Finally, it calculates the bounding box of the grid and rounds the coordinates.
}
