% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{extract_daily_data}
\alias{extract_daily_data}
\title{Extract Daily Data from HDF5 File}
\usage{
extract_daily_data(
  download_file_path,
  h5_data,
  blackmarble_variable,
  quality_flags_to_remove
)
}
\arguments{
\item{download_file_path}{A character string specifying the file path.}

\item{h5_data}{The HDF5 file data.}

\item{blackmarble_variable}{A character string specifying the variable name.}

\item{quality_flags_to_remove}{A numeric vector containing quality flag values to be removed from the data.}
}
\value{
A list containing the extracted data and metadata.
}
\description{
This function extracts daily data from an HDF5 file based on the provided file path and variable name.
}
\details{
This function extracts daily data from an HDF5 file for the specified variable.
It retrieves the data and applies quality flag removal if specified.
Additionally, it calculates the bounding box coordinates based on the tile ID extracted from the file path.
}
\references{
Black Marble User Guide - https://viirsland.gsfc.nasa.gov/PDF/BlackMarbleUserGuide_v1.2_20220916.pdf
}
