% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{extract_data_and_metadata_from_hdf5}
\alias{extract_data_and_metadata_from_hdf5}
\title{Extract Data and Metadata from HDF5 File}
\usage{
extract_data_and_metadata_from_hdf5(
  h5_data,
  download_path,
  blackmarble_variable,
  quality_flags_to_remove
)
}
\arguments{
\item{h5_data}{The HDF5 file data.}

\item{blackmarble_variable}{A character string specifying the variable name.}

\item{quality_flags_to_remove}{A numeric vector containing quality flag values to be removed from the data.}

\item{file_path}{A character string specifying the file path.}
}
\value{
A list containing the extracted data and metadata.
}
\description{
This function extracts data and metadata from an HDF5 file.
}
\details{
This function extracts data and metadata from an HDF5 file based on the specified file path and variable name.
Depending on the file path pattern (daily or monthly/annually), it calls different functions to extract the data.
Quality flag values specified in \code{quality_flags_to_remove} are removed from the data.
}
\references{
Black Marble User Guide - https://viirsland.gsfc.nasa.gov/PDF/BlackMarbleUserGuide_v1.2_20220916.pdf
}
