% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarbler.R
\name{get_nasa_token}
\alias{get_nasa_token}
\title{Get a NASA Earthdata bearer token}
\usage{
get_nasa_token(username, password)
}
\arguments{
\item{username}{character. NASA Earthdata username}

\item{password}{character. NASA Earthdata password}
}
\value{
character
}
\description{
Fetch a NASA Earthdata bearer token from using the Earthdata API.
If none exist, this will create one, or if one already exists it will fetch
that one instead.
}
\author{
Simon E. H. Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
