% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{intersect_bm_tiles}
\alias{intersect_bm_tiles}
\title{Intersect Black Marble Tiles with Region of Interest}
\usage{
intersect_bm_tiles(black_marble_tiles_sf, roi_sf)
}
\arguments{
\item{black_marble_tiles_sf}{Spatial object representing Black Marble tiles.}

\item{roi_sf}{Spatial object representing the region of interest.}
}
\value{
Spatial object containing Black Marble tiles intersecting with the region of interest.
}
\description{
This function intersects Black Marble tiles with a region of interest, removing grid along edges to avoid issues with \code{st_intersects}.
}
\examples{
\dontrun{
# Load required libraries
library(sf)

# Create example data
black_marble_tiles_sf <- st_read("path_to_black_marble_tiles_shapefile")
roi_sf <- st_read("path_to_roi_shapefile")

# Intersect Black Marble tiles with region of interest
intersected_tiles <- intersect_bm_tiles(black_marble_tiles_sf, roi_sf)
}
}
