% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{julian_to_month}
\alias{julian_to_month}
\title{Translate Julian Dates to Regular Month Representation}
\usage{
julian_to_month(julian_date)
}
\arguments{
\item{julian_date}{A character string representing the day of the year in Julian format (e.g., "001" for January 1st).}
}
\value{
A number representing the month corresponding to the given Julian date (e.g., 1 for January).
}
\description{
This function translates Julian dates to regular month representations.
}
\examples{
julian_to_month("001")
# [1] 1

julian_to_month("032")
# [1] 2
}
\references{
For more information on the Julian day system, see: https://en.wikipedia.org/wiki/Julian_day
}
