% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{map_black_marble_tiles}
\alias{map_black_marble_tiles}
\title{Map Black Marble Tiles}
\usage{
map_black_marble_tiles()
}
\value{
An interactive Leaflet map object.
}
\description{
Generate an interactive map displaying black marble tiles using Leaflet.
}
\details{
This function requires the \code{leaflet} package for generating interactive maps.

This function generates an interactive map displaying black marble tiles obtained from NASA's Black Marble project using the Leaflet library. The map includes black marble tiles overlaid on a base map, with a marker indicating the center coordinates (latitude and longitude).
}
\examples{
\dontrun{
library(leaflet)
# Generate the map
map_black_marble_tiles()

}
}
