% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{read_black_marble_csv}
\alias{read_black_marble_csv}
\title{Read Black Marble CSV Data}
\usage{
read_black_marble_csv(year, day, product_id)
}
\arguments{
\item{year}{The year of the data.}

\item{day}{The day of the year (1-365 or 1-366 for leap years).}

\item{product_id}{The product ID specifying the type of data to retrieve.}
}
\value{
A data frame containing the Black Marble CSV data for the specified year and day.
}
\description{
Reads Black Marble CSV data for a specific year and day.
}
\details{
This function reads Black Marble CSV data from the NASA LADS website for the given \code{year},
\code{day}, and \code{product_id}. It constructs the URL based on the provided parameters and attempts
to read the CSV file using \code{readr::read_csv}. If successful, it adds columns for \code{year} and \code{day}
to the data frame. If an error occurs during the reading process, it returns an empty data frame and issues a warning.
Additionally, to avoid overloading the server with rapid requests, the function includes a small delay (0.1 seconds) between
consecutive requests using \code{Sys.sleep(0.1)}.
}
\examples{
\dontrun{
# Read Black Marble CSV data for the year 2023, day 150, and product ID "VIIRS_SNPP_CorrectedReflectance_BandM3"
data <- read_black_marble_csv(2023, 150, "VIIRS_SNPP_CorrectedReflectance_BandM3")
}
}
