% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{retrieve_and_process_nightlight_data}
\alias{retrieve_and_process_nightlight_data}
\title{Retrieve and Process Nightlight Data}
\usage{
retrieve_and_process_nightlight_data(
  roi_sf,
  product_id,
  date,
  bearer,
  blackmarble_variable,
  quality_flags_to_remove,
  check_all_tiles_exist = TRUE,
  quiet = FALSE,
  temp_dir
)
}
\arguments{
\item{roi_sf}{An sf object representing the region of interest.}

\item{product_id}{A character string specifying the product ID.}

\item{date}{A character string representing the date.}

\item{bearer}{A character string representing the authorization bearer token.}

\item{blackmarble_variable}{A character string specifying the blackmarble_variable to extract.}

\item{quality_flags_to_remove}{A numeric vector containing quality flag values to be removed from the data (optional).}

\item{check_all_tiles_exist}{A logical value indicating whether to check if all satellite imagery tiles for the location exist (default is TRUE).}

\item{quiet}{A logical value indicating whether to suppress progress messages (default is FALSE).}

\item{temp_dir}{A character string representing the temporary directory path.}
}
\value{
A raster object containing the processed nighttime light data.
}
\description{
Retrieves and processes nighttime light raster data from the Black Marble dataset.
}
\details{
This function retrieves and processes nighttime light raster data from the Black Marble dataset.
It downloads the data for the specified region of interest and date, removes any quality-flagged
pixels as specified, and mosaics the tiles together if necessary. It then crops the raster to the
region of interest.
}
