% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarbler.R
\name{bm_extract}
\alias{bm_extract}
\title{Extract and process raster data for individual dates}
\usage{
bm_extract(
  roi_sf,
  product_id,
  date,
  bearer,
  aggregation_fun = c("mean"),
  add_n_pixels = TRUE,
  variable = NULL,
  quality_flags_to_remove = NULL,
  check_all_tiles_exist = TRUE,
  interpol_na = FALSE,
  output_location_type = "memory",
  file_dir = NULL,
  file_prefix = NULL,
  file_skip_if_exists = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{roi_sf}{The spatial features representing the regions of interest.}

\item{product_id}{The product ID.}

\item{date}{The date.}

\item{bearer}{The bearer.}

\item{aggregation_fun}{The function to aggregate data.}

\item{add_n_pixels}{Logical indicating whether to compute additional pixel information.}

\item{variable}{The variable.}

\item{quality_flags_to_remove}{The quality flags to remove.}

\item{check_all_tiles_exist}{Logical indicating whether to check if all tiles exist.}

\item{interpol_na}{Logical indicating whether to interpolate missing values.}

\item{output_location_type}{The type of output location.}

\item{file_dir}{The directory to save files.}

\item{file_prefix}{The prefix for file names.}

\item{file_skip_if_exists}{Logical indicating whether to skip saving if files exist.}

\item{quiet}{Logical indicating whether to suppress messages.}

\item{...}{Additional arguments passed to helper functions.}
}
\value{
A data frame containing the extracted and processed raster data.
}
\description{
This function extracts and processes raster data for individual dates.
}
\examples{
bm_extract(roi_sf = my_roi_sf, product_id = "my_product", date = "2024-04-02", bearer = "my_bearer")
}
