% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{count_n_obs}
\alias{count_n_obs}
\title{Count Observations for Exact Extract}
\usage{
count_n_obs(values)
}
\arguments{
\item{values}{A data frame containing the values.}

\item{coverage_fraction}{A numeric value representing the coverage fraction.}
}
\value{
A data frame with the count of non-NA pixels and total pixels for each variable.
}
\description{
Counts observations for each variable, considering coverage fraction, for use in exact_extract.
}
\details{
This function counts observations for each variable in the provided data frame, considering coverage fraction,
for use in exact_extract. It calculates the number of non-NA pixels and total pixels for each variable. The coverage_fraction
parameter is included in the function's signature but currently not used in the calculation.
}
