% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{define_blackmarble_variable}
\alias{define_blackmarble_variable}
\title{Define Black Marble Variable}
\usage{
define_blackmarble_variable(variable, product_id)
}
\arguments{
\item{variable}{A character string specifying the variable to define.}

\item{product_id}{A character string representing the product ID.}
}
\value{
A character string representing the defined variable.
}
\description{
Defines the variable based on the Black Marble product ID if it is NULL.
}
