% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{extract_monthly_data}
\alias{extract_monthly_data}
\title{Extract Monthly Data from HDF5 File}
\usage{
extract_monthly_data(
  download_file_path,
  h5_data,
  blackmarble_variable,
  quality_flags_to_remove
)
}
\description{
Extract Monthly Data from HDF5 File
}
