% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmarble_utils.R
\name{extract_monthly_data}
\alias{extract_monthly_data}
\title{Extract Monthly Data from HDF5 File}
\usage{
extract_monthly_data(
  download_file_path,
  h5_data,
  blackmarble_variable,
  quality_flags_to_remove
)
}
\arguments{
\item{download_file_path}{Path to the downloaded HDF5 file.}

\item{h5_data}{The HDF5 data.}

\item{blackmarble_variable}{The variable to extract from the HDF5 data.}

\item{quality_flags_to_remove}{Quality flag values to set to NA.}
}
\value{
A list containing the extracted data and bounding box coordinates.
}
\description{
This function extracts monthly data from an HDF5 file.
}
\examples{
# Extract monthly data
data <- extract_monthly_data(download_file_path, h5_data, blackmarble_variable, quality_flags_to_remove)

}
